+++ js/utils.js
@@ -1,0 +1,15 @@
+function getRandomNumber(min, max) {
+    const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+    const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+    const result = Math.random() * (upper - lower + 1) + lower;
+    return Math.floor(result);
+  }
+  
+  function getRandomNumberPoint(min, max, digits = 4) {
+    const lower = Math.min(Math.abs(min), Math.abs(max));
+    const upper = Math.max(Math.abs(min), Math.abs(max));
+    const result = Math.random() * (upper - lower) + lower;
+    return +result.toFixed(digits);
+  }
+
+export {getRandomNumber, getRandomNumberPoint};
